cambios

+++ FarmaciaElPalenque/Models/Productos.cs
@@ -4,7 +4,7 @@
     {
         public int id { get; set; }
         public string? nombre { get; set; }
-        public decimal precio { get; set; }
+        public int precio { get; set; }
         public int categoriaId { get; set; }
         public Categoria? Categoria { get; set; }
         public string? imagenUrl { get; set; }
